+++ test/task/MonitorTask.test.ts
@@ -972,7 +972,8 @@ describe('MonitorTask', () => {
 
       it('does not degrade video codec when metrics are normal with software encoder', () => {
         connectionHealthData.setIsVideoEncoderHardware(false);
-        connectionHealthData.setVideoEncodingTimeInMs(20);
+        connectionHealthData.setVideoEncodingTimeInMs(300);
+        connectionHealthData.setVideoEncodingTimePerFrameInMs(10);
         connectionHealthData.setCpuLimitationDuration(0);
         connectionHealthData.setVideoInputFps(15);
         connectionHealthData.setVideoEncodeFps(15);
@@ -987,7 +988,8 @@ describe('MonitorTask', () => {
 
       it('does not degrade video codec when metrics are normal with hardware encoder', () => {
         connectionHealthData.setIsVideoEncoderHardware(true);
-        connectionHealthData.setVideoEncodingTimeInMs(30);
+        connectionHealthData.setVideoEncodingTimeInMs(450);
+        connectionHealthData.setVideoEncodingTimePerFrameInMs(10);
         connectionHealthData.setCpuLimitationDuration(0);
         connectionHealthData.setVideoInputFps(15);
         connectionHealthData.setVideoEncodeFps(15);
@@ -1002,8 +1004,11 @@ describe('MonitorTask', () => {
 
       it('does degrade video codec when software encoding CPU is constantly high', () => {
         connectionHealthData.setIsVideoEncoderHardware(false);
-        connectionHealthData.setVideoEncodingTimeInMs(600);
+        connectionHealthData.setVideoEncodingTimeInMs(900);
+        connectionHealthData.setVideoEncodingTimePerFrameInMs(60);
         connectionHealthData.setCpuLimitationDuration(0);
+        connectionHealthData.setVideoInputFps(15);
+        connectionHealthData.setVideoEncodeFps(15);
         for (let i = 0; i < 15; i++) {
           task.connectionHealthDidChange(connectionHealthData);
         }
@@ -1015,6 +1020,7 @@ describe('MonitorTask', () => {
 
       it('does degrade video codec when video quality is limited due to CPU', () => {
         connectionHealthData.setIsVideoEncoderHardware(false);
+        connectionHealthData.setVideoEncodingTimePerFrameInMs(0);
         connectionHealthData.setVideoEncodingTimeInMs(0);
         connectionHealthData.setCpuLimitationDuration(1);
         for (let i = 0; i < 15; i++) {
@@ -1028,6 +1034,7 @@ describe('MonitorTask', () => {
 
       it('does degrade video codec when video encoding fails', () => {
         connectionHealthData.setIsVideoEncoderHardware(true);
+        connectionHealthData.setVideoEncodingTimePerFrameInMs(0);
         connectionHealthData.setVideoEncodingTimeInMs(0);
         connectionHealthData.setCpuLimitationDuration(1);
         connectionHealthData.setVideoInputFps(15);